requirements.txt
Django==2.2.7
psycopg2-binary==2.7.7
psycopg2    ==2.7.7
dj-database-url==0.5.0    
djangorestframework==3.10.0
djangorestframework-simplejwt==3.3
django-cors-headers==2.4.0
django-filter==2.0.0
django-request-logging==0.6.6
django-request-profiler==0.13
django-storages==1.7.1
djangorestframework-csv==2.1.0
djangorestframework-gis==0.14cd
requests==2.22.0
requests-oauthlib==1.0.0
requests-toolbelt==0.7.0
whitenoise==3.3.1
Pillow==5.4.1